move SDK package as children of data package
