+++ app/src/main/java/org/eyeseetea/malariacare/LoginActivity.java
@@ -44,7 +44,7 @@ import android.widget.TextView;
 import org.eyeseetea.malariacare.database.utils.PreferencesState;
 import org.eyeseetea.malariacare.domain.usecase.LoginUseCase;
 import org.eyeseetea.malariacare.network.ServerAPIController;
-import org.eyeseetea.malariacare.sdk.SdkLoginController;
+import org.eyeseetea.malariacare.data.sdk.SdkLoginController;
 import org.eyeseetea.malariacare.strategies.LoginActivityStrategy;
 import org.eyeseetea.malariacare.utils.Utils;
 import org.hisp.dhis.client.sdk.ui.activities.AbsLoginActivity;
