+++ app/src/main/java/org/eyeseetea/malariacare/data/sdk/SdkPullController.java
@@ -1,5 +1,88 @@
-package org.eyeseetea.malariacare.sdk;
+package org.eyeseetea.malariacare.data.sdk;
 
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
+import static org.eyeseetea.malariacare.data.sdk.SdkController.postFinish;
 import static org.hisp.dhis.client.sdk.models.program.ProgramType.WITHOUT_REGISTRATION;
 
 import android.content.Context;
