+++ app/src/main/java/org/eyeseetea/malariacare/database/iomodules/dhis/importer/models/ProgramExtended.java
@@ -29,10 +29,8 @@ import android.util.Log;
 import com.raizlabs.android.dbflow.sql.language.Join;
 import com.raizlabs.android.dbflow.sql.language.Select;
 
-import org.eyeseetea.malariacare.database.iomodules.dhis.importer.IConvertFromSDKVisitor;
-import org.eyeseetea.malariacare.database.iomodules.dhis.importer.VisitableFromSDK;
 import org.eyeseetea.malariacare.database.model.Program;
-import org.eyeseetea.malariacare.sdk.SdkQueries;
+import org.eyeseetea.malariacare.data.sdk.SdkQueries;
 import org.hisp.dhis.client.sdk.android.api.persistence.flow.AttributeFlow;
 import org.hisp.dhis.client.sdk.android.api.persistence.flow.AttributeFlow_Table;
 import org.hisp.dhis.client.sdk.android.api.persistence.flow.ProgramAttributeValueFlow;
