+++ app/src/main/java/org/eyeseetea/malariacare/layout/listeners/PullRequiredOnPreferenceChangeListener.java
@@ -6,7 +6,7 @@ import android.preference.Preference;
 import org.eyeseetea.malariacare.R;
 import org.eyeseetea.malariacare.database.utils.PreferencesState;
 import org.eyeseetea.malariacare.network.ServerAPIController;
-import org.eyeseetea.malariacare.sdk.SdkLoginController;
+import org.eyeseetea.malariacare.data.sdk.SdkLoginController;
 
 public class PullRequiredOnPreferenceChangeListener implements
         Preference.OnPreferenceChangeListener {
